# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TodoPay API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  loginId: string;
  password: string;
}

export interface AdminUser {
  id: number;
  loginId: string;
  name: string;
  /** superadmin | hq | distributor | agency | store */
  role: string;
  /** readonly | admin | finance */
  permission: string;
  isActive: boolean;
  useOtp: boolean;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  parentName?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: AdminUser;
  token: string;
}

export interface AdminUserList {
  items: AdminUser[];
  total: number;
}

export interface AdminUserInput {
  loginId: string;
  password: string;
  name: string;
  role: string;
  permission: string;
  /** @nullable */
  parentId?: number | null;
  useOtp?: boolean;
}

export interface AdminUserUpdate {
  name?: string;
  isActive?: boolean;
  permission?: string;
  useOtp?: boolean;
}

export interface ResetPasswordInput {
  newPassword: string;
}

export interface UserPermissionUpdate {
  permission: string;
}

export interface Member {
  id: number;
  loginId: string;
  name: string;
  phone: string;
  email: string;
  storeCode: string;
  storeName: string;
  isActive: boolean;
  /** @nullable */
  virtualAccountNumber?: string | null;
  /** @nullable */
  virtualAccountBank?: string | null;
  createdAt: string;
}

export interface MemberList {
  items: Member[];
  total: number;
}

export interface MemberInput {
  loginId: string;
  password: string;
  name: string;
  phone: string;
  email: string;
  storeCode: string;
}

export interface MemberUpdate {
  name?: string;
  phone?: string;
  email?: string;
  isActive?: boolean;
}

export interface StatusUpdate {
  isActive: boolean;
}

export interface Buyer {
  id: number;
  name: string;
  loginId: string;
  phone: string;
  birthdate: string;
  isVerified: boolean;
  virtualAccountNumber: string;
  virtualAccountBank: string;
  /** active | revoked */
  virtualAccountStatus: string;
  /** @nullable */
  withdrawalAccount?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface BuyerList {
  items: Buyer[];
  total: number;
}

export interface BuyerInput {
  name: string;
  loginId: string;
  password: string;
  phone: string;
  birthdate: string;
}

export interface RegisterLink {
  url: string;
}

export interface VirtualAccount {
  id: number;
  accountNumber: string;
  bankName: string;
  /** active | revoked */
  status: string;
  memberId: number;
  memberName: string;
  balance?: number;
  createdAt: string;
}

export interface VirtualAccountList {
  items: VirtualAccount[];
  total: number;
}

export interface Withdrawal {
  id: number;
  trackingNumber: string;
  amount: number;
  fee: number;
  totalAmount: number;
  /** pending | approved | rejected */
  approvalStatus: string;
  /** unpaid | paid */
  withdrawalStatus: string;
  accountNumber: string;
  accountBank: string;
  accountHolder: string;
  /** @nullable */
  rejectReason?: string | null;
  /** @nullable */
  memberName?: string | null;
  /** @nullable */
  storeName?: string | null;
  createdAt: string;
}

export interface WithdrawalList {
  items: Withdrawal[];
  total: number;
  totalAmount: number;
  totalFee: number;
}

export interface WithdrawalInput {
  amount: number;
  accountNumber: string;
  accountBank: string;
  accountHolder: string;
}

export interface RejectInput {
  reason: string;
}

export interface WithdrawalSummary {
  pendingCount: number;
  pendingAmount: number;
  approvedCount: number;
  approvedAmount: number;
  todayWithdrawn: number;
}

export interface Transaction {
  id: number;
  /** deposit | withdrawal */
  type: string;
  originalAmount: number;
  amount: number;
  fee: number;
  /** success | failed | pending */
  status: string;
  fromAccount: string;
  toAccount: string;
  trackingNumber: string;
  pgTransactionId: string;
  /** @nullable */
  memberName?: string | null;
  createdAt: string;
}

export interface TransactionList {
  items: Transaction[];
  total: number;
}

export interface BalanceRecord {
  id: number;
  /** in | out */
  direction: string;
  /** withdrawal | payment */
  category: string;
  amount: number;
  balance: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface BalanceRecordList {
  items: BalanceRecord[];
  total: number;
}

export interface BalanceSummary {
  balance: number;
  pendingAmount: number;
}

export interface FeeConfig {
  id: number;
  userId: number;
  userName: string;
  role: string;
  depositFee: number;
  withdrawalFee: number;
  createdAt: string;
}

export interface FeeConfigInput {
  userId: number;
  depositFee: number;
  withdrawalFee: number;
}

export interface FeeConfigUpdate {
  depositFee?: number;
  withdrawalFee?: number;
}

export interface DailyStatistic {
  date: string;
  depositCount: number;
  depositAmount: number;
  withdrawalCount: number;
  withdrawalAmount: number;
  feeAmount: number;
  netAmount: number;
}

export interface StatisticsOverview {
  todayDeposit: number;
  todayWithdrawal: number;
  todayFee: number;
  monthDeposit: number;
  monthWithdrawal: number;
  totalMembers: number;
  activeVirtualAccounts: number;
  pendingWithdrawals: number;
}

export interface Notice {
  id: number;
  title: string;
  content: string;
  isPinned: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface NoticeList {
  items: Notice[];
  total: number;
}

export interface NoticeInput {
  title: string;
  content: string;
  isPinned?: boolean;
}

export interface NoticeUpdate {
  title?: string;
  content?: string;
  isPinned?: boolean;
}

export interface OtpSettings {
  useOtpForDeposit: boolean;
  useOtpForWithdrawal: boolean;
  /** @nullable */
  otpSecret?: string | null;
}

export interface OtpSettingsUpdate {
  useOtpForDeposit?: boolean;
  useOtpForWithdrawal?: boolean;
}

export type ListUsersParams = {
  role?: string;
  /**
   * @nullable
   */
  parentId?: number | null;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListMembersParams = {
  storeCode?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListBuyersParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  page?: number;
  limit?: number;
};

export type ListVirtualAccountsParams = {
  /**
   * @nullable
   */
  memberId?: number | null;
  status?: string;
  page?: number;
  limit?: number;
};

export type ListWithdrawalsParams = {
  startDate?: string;
  endDate?: string;
  search?: string;
  approvalStatus?: string;
  withdrawalStatus?: string;
  page?: number;
  limit?: number;
};

export type ListTransactionsParams = {
  startDate?: string;
  endDate?: string;
  type?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListBalanceRecordsParams = {
  startDate?: string;
  endDate?: string;
  type?: string;
  page?: number;
  limit?: number;
};

export type ListFeesParams = {
  /**
   * @nullable
   */
  userId?: number | null;
};

export type GetDailyStatisticsParams = {
  startDate?: string;
  endDate?: string;
};

export type ListNoticesParams = {
  page?: number;
  limit?: number;
};