# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TodoPay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members
      parameters:
        - name: storeCode
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberList"
    post:
      operationId: createMember
      tags: [members]
      summary: Register member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/status:
    patch:
      operationId: updateMemberStatus
      tags: [members]
      summary: Update member status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Status updated

  # Buyers (구매자)
  /buyers:
    get:
      operationId: listBuyers
      tags: [buyers]
      summary: List buyers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of buyers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyerList"
    post:
      operationId: createBuyer
      tags: [buyers]
      summary: Register buyer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Buyer"
  /buyers/{id}:
    get:
      operationId: getBuyer
      tags: [buyers]
      summary: Get buyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Buyer"
    delete:
      operationId: deleteBuyer
      tags: [buyers]
      summary: Delete buyer (revoke virtual account)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /buyers/{id}/virtual-account:
    post:
      operationId: reissueBuyerVirtualAccount
      tags: [buyers]
      summary: Reissue virtual account for buyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reissued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Buyer"
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee configurations
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Fee configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeConfig"
    post:
      operationId: createFeeConfig
      tags: [fees]
      summary: Create fee configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeConfigInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeConfig"
  /fees/{id}:
    patch:
      operationId: updateFeeConfig
      tags: [fees]
      summary: Update fee configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeConfigUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeConfig"

  # Statistics (일자별 통계)
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [loginId, password]
      properties:
        loginId:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        token:
          type: string

    # Admin Users
    AdminUser:
      type: object
      required: [id, loginId, name, role, permission, isActive, useOtp, createdAt]
      properties:
        id:
          type: integer
        loginId:
          type: string
        name:
          type: string
        role:
          type: string
          description: "superadmin | hq | distributor | agency | store"
        permission:
          type: string
          description: "readonly | admin | finance"
        isActive:
          type: boolean
        useOtp:
          type: boolean
        parentId:
          type: ["integer", "null"]
        parentName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Member:
      type: object
      required: [id, loginId, name, phone, email, storeCode, storeName, isActive, createdAt]
      properties:
        id:
          type: integer
        loginId:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        storeCode:
          type: string
        storeName:
          type: string
        isActive:
          type: boolean
        virtualAccountNumber:
          type: ["string", "null"]
        virtualAccountBank:
          type: ["string", "null"]
        createdAt:
          type: string
    MemberList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total:
          type: integer
    MemberInput:
      type: object
      required: [loginId, password, name, phone, email, storeCode]
      properties:
        loginId:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        storeCode:
          type: string
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        isActive:
          type: boolean
    StatusUpdate:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean

    # Buyers
    Buyer:
      type: object
      required: [id, name, loginId, phone, birthdate, isVerified, virtualAccountNumber, virtualAccountBank, virtualAccountStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        loginId:
          type: string
        phone:
          type: string
        birthdate:
          type: string
        isVerified:
          type: boolean
        virtualAccountNumber:
          type: string
        virtualAccountBank:
          type: string
        virtualAccountStatus:
          type: string
          description: "active | revoked"
        withdrawalAccount:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BuyerList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Buyer"
        total:
          type: integer
    BuyerInput:
      type: object
      required: [name, loginId, password, phone, birthdate]
      properties:
        name:
          type: string
        loginId:
          type: string
        password:
          type: string
        phone:
          type: string
        birthdate:
          type: string
...
    FeeConfig:
      type: object
      required: [id, userId, userName, role, depositFee, withdrawalFee, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        role:
          type: string
        depositFee:
          type: number
        withdrawalFee:
          type: number
        createdAt:
          type: string
    FeeConfigInput:
      type: object
      required: [userId, depositFee, withdrawalFee]
      properties:
        userId:
          type: integer
        depositFee:
          type: number
        withdrawalFee:
          type: number
    FeeConfigUpdate:
      type: object
      properties:
        depositFee:
          type: number
        withdrawalFee:
          type: number

    # Statistics
...