# lib/api-zod/src/generated/types/buyerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TodoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BuyerInput {
  name: string;
  loginId: string;
  password: string;
  phone: string;
  birthdate: string;
}