# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TodoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  name?: string;
  phone?: string;
  email?: string;
  isActive?: boolean;
}